Added movemouse fn and example
The added function takes in 6 parameters: Robot object (for mouse movement), initial position x, initial position y, displacement x, displacement y, speed (1 is a good amount, 3 is great for testing). The example program shows a working example of a mouse moving in a <u>M</u> formation by looped calls to the above function and clicking a button when done.